1) фикстура вынесена в отдельный файл conftest; 2) фикструра убрана из тестов; 3) создан тест test_del_group
