+++ conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from fixture.application import Application
+
+
+@pytest.fixture
+def app(request):
+    fixture = Application()
+    request.addfinalizer(fixture.destroy)
+    return fixture
