+++ test/test_add_contact.py
@@ -1,14 +1,5 @@
 # -*- coding: utf-8 -*-
-import pytest
 from model.contact_param import ContactParam
-from fixture.application import Application
-
-
-@pytest.fixture
-def app(request):    # грубо app = Application()
-    fixture = Application()    # фикстура - экземпляр класса Application, созданный перед запуском теста.
-    request.addfinalizer(fixture.destroy)
-    return fixture
 
 
 def test_add_contact(app):
