+++ test/test_add_group.py
@@ -1,23 +1,14 @@
 # -*- coding: utf-8 -*-
-import pytest
 from model.group_param import GroupParam
-from fixture.application import Application
-
-
-@pytest.fixture
-def app(request):
-    fixture = Application()
-    request.addfinalizer(fixture.destroy)
-    return fixture
 
 
 def test_add_group(app):
     app.session.login(user_name="admin", password="secret")    # .session т.к. login перенесли в fixture.session
-    app.group.create(GroupParam(group_name="Meeting", header="Hello", footer="Bye"))
+    app.group.create(GroupParam(group_name="Meeting1", header="Hello1", footer="Bye1"))
     app.session.logout()
 
 
 def test_add_empty_group(app):
     app.session.login(user_name="admin", password="secret")
-    app.group.create(GroupParam(group_name="", header="", footer=""))
+    app.group.create(GroupParam(group_name="1", header="2", footer="3"))
     app.session.logout()
