+++ test/test_del_group.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+def test_delete_first_group(app):
+    app.session.login(user_name="admin", password="secret")    # .session т.к. login перенесли в fixture.session
+    app.group.delete_first_group()
+    app.session.logout()
